requirements.txt
aiofiles==24.1.0 ; python_version >= "3.12" and python_version < "4.0"
aiohappyeyeballs==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.10.5 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
asn1crypto==1.5.1 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4"
boto3==1.35.24 ; python_version >= "3.12" and python_version < "4.0"
botocore==1.35.24 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4"
cloud-sql-python-connector==1.12.1 ; python_version >= "3.12" and python_version < "4.0"
cloudevents==1.11.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4" and platform_system == "Windows"
cryptography==42.0.8 ; python_version >= "3.12" and python_version < "4.0"
datetime==5.5 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.14 ; python_version >= "3.12" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
flask==3.0.3 ; python_version >= "3.12" and python_version < "4"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
functions-framework==3.8.1 ; python_version >= "3.12" and python_version < "4"
google-api-core==2.20.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.20.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.4.5 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.11.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4" and platform_system != "Windows"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.4.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4"
jmespath==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4"
multidict==6.1.0 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.27.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4"
pg8000==1.31.2 ; python_version >= "3.12" and python_version < "4.0"
pillow-heif==0.15.0 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.12" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
s3transfer==0.10.2 ; python_version >= "3.12" and python_version < "4.0"
scramp==1.4.5 ; python_version >= "3.12" and python_version < "4.0"
setuptools==75.1.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
slack-bolt==1.20.1 ; python_version >= "3.12" and python_version < "4.0"
slack-sdk==3.33.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy-utils==0.41.2 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.35 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
watchdog==5.0.2 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.0.4 ; python_version >= "3.12" and python_version < "4"
wrapt==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.11.1 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.20.2 ; python_version >= "3.12" and python_version < "4.0"
zope-interface==7.0.3 ; python_version >= "3.12" and python_version < "4.0"
